requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.22.0
google-api-python-client>=2.0.0
openai==1.54.4
python-dotenv==1.0.1